[#27] Check Exercise Files Before Starting Selenium
Modify the pluraexercise script so that it removes and logs any courses
that we have already downloaded the exercises for. If there are any
remaining exercise files to download, we start the Chrome driver and
fetch them.

This prevents the starting of Chrome & the login action if we've already
downloaded every exercise file in the courselist.

Closes #27