[#36] Increase Safety of Moving Course Files
Modify the move_content function so that instead of completely replacing
the target directory with the source directory, it moves the files
inside the source directory and then removes the directory. This
prevents file loss when a subset of files have been downloaded after
another completed subset.

Do not delete the INPROGRESSPATH afterwards, in case there are other
course folders inside.

Fixes #36